undo change to main news manifest
i still want the new release announcement to be shown there. the actual page
links to the announcement text under docs/release/ now. under normal situations
i don't want release announcements under news/, but this was a very special
release because it was the first release in 5 years

i will leave it there