nuke outdated distro-specific FDE guides
i do not intend to re-add them. these guides should be on the wiki pages of
each project website e.g. parabola wiki

it's out of scope for libreboot. the debian guide will remain, for now, but as
soon as it becomes broken or no longer fit for purpose, it will simply be
deleted. ideally, the pages on docs/gnulinux/ should not be distro-specific