delete libreboot.css. embed it into template.include
The CSS is light enough now that, in terms of compression, it makes more sense
to merge it in each HTML page. It's small enough to make virtually no
difference, and this cuts down on HTTP requests.

This change will save literally seconds of time, for those who are on high
latency internet services (satellite based internet services, slow tor relays,
etc).