kludge: rename sl.png to favicon.ico
and host it on libreboot.org/ instead of av.libreboot.org/

the PNG "light" libreboot logo is well-optimized, and still opens properly in
most browsers even if named favicon.ico. the icon is going to be loaded on
pages anyway. having a separate favicon.ico adds an extra http request, which
causes seconds of lag on some connections

so make sl.png the favicon, and use that in the img tag on pages