Generate a tarball of the website
Sourcehut has a way to setup a static website by uploading a tarball
of the content[1].

Even if we don't end up using Sourcehut, generating a tarball of the
website enables more easy deployments.

Note that we didn't touch to the website code yet, so it still uses
the old URLs, the old image locations, etc.

After creating a token (documentation[1]), the website can then be
uploaded with the following command:
curl \
	--oauth2-bearer "<token>" \
	-Fcontent=@website.tar.gz \
	https://pages.sr.ht/publish/gnutoo.srht.site

[1]https://srht.site/quickstart

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>