Fix first Libreboot release date
The first Libreboot release was the 12th December 2013[1].

This is also correlated with the fact that I moved to France around
summer 2013, and around December, I gave Leah my build scripts (under
a free license) and I answered all her questions in order to help her
getting the first RYF computer certified. That then lead to the
creation of Libreboot. So the first release cound't have happened in
2009.

[1]https://libreboot.org/news/libreboot20131212.html

Reported-by: f_ (Ferass El Hafidi) on #libreboot on Liberachat.
Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>
Acked-by: Adrien 'neox' Bourmault <neox@a-lec.org>