Continue Libreboot under the GNU project
This commit is based on the 20220710 tag from Libreboot.

To our knowledge, 20220710 is the last really libre Libreboot
release as the next releases from libreboot.org has nonfree
software (like nonfree microcode updates).

Because of that we've stepped forward to stand up for freedom,
and we started maintaining our own version of Libreboot that
didn't include nonfree software.

To make sure that our version remains free and continue to be
maintained over time, we chose to now do this as part of the
GNU project.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>
GNUtoo: commit message
Acked-by: Adrien 'neox' Bourmault <neox@gnu.org>