coreboot/fam15h: update code base, deblob, unset CONFIG_STM (see bug #64535)
This commit updates the coreboot code base from release 4.11 to 4.11_branch for kgpe-d16,
kcma-d8, kfsn4-dre and addresses one new blob related to this update.

The main reason to update the codebase is to prevent a bug with RAM initialization
that occured with coreboot 4.11 and raised the following critical error:

	fam15_receiver_enable_training_seed: using seed: 0054
	fam15_receiver_enable_training_seed: using seed: 0054
	TrainRcvrEn: Status 2005
	TrainRcvrEn: ErrStatus 4000
	TrainRcvrEn: ErrCode 0
	TrainRcvrEn: Done

	TrainDQSReceiverEnCyc_D_Fam15: lane 0 failed to train!  Training for receiver 2 on DCT 0 aborted
	TrainDQSReceiverEnCyc: Status 2205
	TrainDQSReceiverEnCyc: TrainErrors 44000
	TrainDQSReceiverEnCyc: ErrStatus 44000
	TrainDQSReceiverEnCyc: ErrCode 0
	TrainDQSReceiverEnCyc: Done

	TrainDQSReceiverEnCyc: Status 2005
	TrainDQSReceiverEnCyc: TrainErrors 4000
	TrainDQSReceiverEnCyc: ErrStatus 4000
	TrainDQSReceiverEnCyc: ErrCode 0
	TrainDQSReceiverEnCyc: Done

	DIMM training FAILED!  Restarting system...soft_reset() called!

This coreboot revision also correct some bugs with SMM, SMBIOS, IPMI and BMC.

Some new values in coreboot configuration make coreboot first build stop to prompt
users and forcing them to choose an option to continue:
	- CONFIG_STM
	- CONFIG_DEBUG_IPMI
	- CONFIG_VENDOR_VIA
	- CONFIG_SOUTHBRIDGE_AMD_CIMX_SB900
	- CONFIG_IPMI_FRU_SINGLE_RW_SZ
	- CONFIG_IPMI_KCS_TIMEOUT_MS

A bug has been opened about CONFIG_STM on our bug tracker [1], and we decided,
for now, to unset this option explicitely.

So in this commit we just regenerated configurations for each fam15h board via
coreboot build prompts and copied the resulting configurations in the configuration
folder and that results in the following:
	- unset CONFIG_STM
	- unset CONFIG_DEBUG_IPMI
	- unset CONFIG_VENDOR_VIA
	- unset CONFIG_SOUTHBRIDGE_AMD_CIMX_SB900
	- set CONFIG_IPMI_FRU_SINGLE_RW_SZ=16
	- set CONFIG_IPMI_KCS_TIMEOUT_MS=5000

[1]https://savannah.gnu.org/bugs/?64535

Signed-off-by: Adrien 'neox' Bourmault <neox@gnu.org>
GNUtoo: split commit into "don't build ada toolchain for generic platforms"
Acked-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>
Acked-by: Adrien 'neox' Bourmault <neox@gnu.org>