Add script to convert markdown files to texi
We can easily convert markdown files with pandoc:
    pandoc --from markdown --to texinfo index.md -o index.texi
    rm -f index.md

And since our modified version of Untitled has support for texi, it
ends up generating html files.

But then we end up with a problem in the title. For instance
index.texi generates a web page with "What is this?
{#What-is-this_003f}" instead of the expected "Free your BIOS today!"
title.

This can be fixed manualy by doing the following change to index.texi:
     @node Top
    -@top Free your BIOS today!
    +@top Top
    +
    +@node
    +@subsection Free your BIOS today!

But there are more than 80 markdown files, so that would require to do
that 80 times manually.

So this quick and dirty script can be used to do that automatically
instead.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>