Fix images path for web usage
We don't plan to use an external website just for the images, so
we need a local path for the images.

Since the images are in www/ in the images repositories, it's easier
to expect the images in img/www to enable easy deployment by users.

The path of the images aren't relative to the page. So we need to use
a kind of absolute path for it to work.

We have the choice of not using any domain (/img/www/) or using a
specific domain (DOMAIN/img/www/). We use the former as the later
doesn't work offline and cannot use the local images.

However the downside is that we need a local web server for it to work
as simply opening the html files in a browser won't work unless the
user puts the images inside /img/www inside the root filesystem.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>