website-build: build.sh: switch to GNU Boot repositories.
Since the website and the images were merged in GNU boot respectively
in site/ and www/, we can now modify GNU boot to only be cloned once
to build the website.

The image path was also modified to make images work both in the local
served website and at gnu.org and also to minimize the differences
between the test and deployment setups.

Libreboot.at also hosted a copy of Untitled to make it simple to
control the revision being used.

But in the longer term, it would be a bad idea to do that again for
GNU boot since there are plans to convert the website and/or
documentation to the texinfo format, and so at some point Untitled
will stop being used, and we don't want to end up with a git
repository that needs to be kept for historic reasons but that is
unused.

Since the website isn't ready we also have a redirection to Savannah
to make sure that visitors are redirected to something meaningful.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>
Neox: Suggested the split in smaller commits and various ways to do
    that (including some that were used).
Acked-by: Adrien 'neox' Bourmault <neox@gnu.org>