website-build: Makefile: website.tar.gz: tar: remove duplicated files.
Since the commit 02d03ecd536635e2e672cf8e1264ca8e9d2f8998
("website-build: build.sh: Start switching to GNU Boot."), we ended up
with twice the images.

The website and the images were merged in GNU boot respectively in
site/ and www/.

The git repository address of both the website and the images were
then updated to point to the GNU boot repository.

But when using the GNU boot git repository for both the website and
the images, the images were included once from the website repository
in untitled/www/lbwww/www and once from the images repository in
untitled/www/lbwww/site/img.

And since the tar command was using untitled/www/, it included the
same images twice.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>