website-build: Use rsync for deploying the website.
The website and documentation originally comes from Libreboot and
Libreboot didn't have separate website and documentation.

And we need to reuse it because without it, GNU Boot would be almost
useless as without documentation most users would not be able to
install it.

The website is about 900 files and consist in about 27 MiB of
pictures.

Given that the website is not ready yet and that we are looking for
contributions to help us fixing it, we will end up having to update it
often, and ideally in a timely manner as well not to discourage
contributions.

So we needed some way to deployment the website with very few
commands.

GNU typically use CVS for deploying the website, but it is harder to
use than rsync for automatizing the deployment of a website.

To do that someone would have to write or adapt code to do the
deployment automatically, but that can potentially be time consuming
to do, especially if it needs to be done efficiently (for instance by
only adding files that changed since the last time) not to load too
much the CVS server.

Because of all that, an rsync access was temporarily setup for us, to
enable us to publish the website now and to be able to fixing this
deployment issue later and/or ask for help on the website to fix it.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>
Acked-by: Adrien 'neox' Bourmault <neox@gnu.org>