website-build: cache tarball builds.
Before we always rebuilt everything to make sure to avoid situations
where newer source files weren't added in the builds.

However this is quite costly as for instance 'make check' ends up
rebuilding the tarball even after running 'make' right before.

Since the Guix revision is fixed and that the build shell script
(build.sh) also hardcodes the revision being used to build the tools,
we can rely on that to only rebuild if they changed.

For external repositories (that are used with ./configure
--with-<repo>-path=PATH [...]) we use a hash of the timestamps of all
the files in these directories. While it's not optimal as any
unrelated change or new file would trigger a rebuild, it at least
improves the situation when people do not use external repositories

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>