Move images in the website directory
Since the website and the images were merged in GNU boot respectively
in site/ and www/, we can now modify GNU boot to only be cloned once
to build the website.

This also requires to move the images inside site/ at the place where
website-build expects them.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>
Neox: Suggested the split from the "website-build: build.sh: switch to
      GNU Boot repositories." commit.
Acked-by: Adrien 'neox' Bourmault <neox@gnu.org>