Move the website to /software/gnuboot/test
The website will be published to https://gnu.org/software/gnuboot. But
for now it is not ready yet to be published as-is because there are no
contribution instructions yet and we also need to agree on what to put
on the main page.

So until that's fixed, it will be published to
https://gnu.org/software/gnuboot/test.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>
Neox: Suggested the split from the "website-build: build.sh: switch to
      GNU Boot repositories." commit.
Acked-by: Adrien 'neox' Bourmault <neox@gnu.org>