README.md: use less harsh language in the project description (bug #64661)
This commit addresses a problem in the wording of our project description
(in the Savannah description and in this README.md).

The project description was re-written from scratch by GNUtoo and this
commits import it unmodified from Savannah[1]. It is also released by
GNUtoo under the following license:
    Copyright (c)  2023  Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>.
    Permission is granted to copy, distribute and/or modify this document
    under the terms of the GNU Free Documentation License, Version 1.3
    or any later version published by the Free Software Foundation;
    with no Invariant Sections, no Front-Cover Texts, and no Back-Cover Texts.
    A copy of the license is included in the section entitled "GNU
    Free Documentation License".

[1]https://savannah.gnu.org/projects/gnuboot

Reported-by: Patrick Georgi <oxygene>
Neox: import, commit message apart from the copyright information.
GNUtoo: copyright information, whitespace fixes.
Signed-off-by: Adrien 'neox' Bourmault <neox@gnu.org>