site: clarify maintainer status.
In Parabola, the libreboot-utils and various u-boot packages needed
Libreboot as it was providing deblobbed source releases of Coreboot
and u-boot.

Because of that and also because I need also deblobbed u-boot for Guix
and Replicant, I ended up being involved in Libreboot.at and then GNU
Boot.

However the Replicant project also needs my help, so when joining
Libreboot.at and then GNU Boot I mostly wanted to have a project I
could rely on and be able to easily send patches to it.

So the idea was that once the project would be in a good shape I would
reconsider if I still needed to be a maintainer and try to see if new
contributors wanted to replace me (to have more than one maintainers)
and then go back sending patches from time to time as a regular
contributor to have more time for other projects.

This is why we used the word "interim" at the beginning: I was not
sure (and still am not sure) to want to stay maintainer of GNU Boot
forever. Adrien however has other plans.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>
Acked-by: Adrien 'neox' Bourmault <neox@gnu.org>