site: send patches: Update the "Name not required" section.
GNU Boot might have to send some of the patches upstream so we can't
have a policy that applies to everything unless we also look into
upstream projects and GNU policies as well.

In addition we might need multiple ways to satisfy upstream
constraints to fit different needs / compromise from people wanting to
contribute anonymously.

And as the problem raised here is more global it is better to have a
more global answers and find ways that also work with upstream projects
as well.

Being a GNU project, GNU boot would also need to look into the GNU
policies on the topic and also try to see if for instance the FSF
copyright assignements could also be used in some cases to fix that
issue.

There are also other things worth looking into as well. For instance
Software Conservancy also handles copyright assignments and Asahi Lina
that contributes anonymously to the Linux kernel while satisfying
their very strict policy about tracking copyright ownership.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>
Acked-by: Adrien 'neox' Bourmault <neox@gnu.org>