Deploy the website on https://gnu.org/software/gnuboot/.
Since we now have a very visible banner that indicates which pages
have been not been reviewed, and contribution instructions on how to
fix that, we are ready to publish the website.

Right now the deploy path of the website isn't configurable as this
would require some form of templating or processing of the markdown
files, so the website was simply moved from
gnu.org/software/gnuboot/test to gnu.org/software/gnuboot.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>
Acked-by: Adrien 'neox' Bourmault <neox@gnu.org>