site: add GNU Boot status page.
Many projects that make hardware work (drivers, etc) have status
pages.

Here GNU Boot is a distribution and with very few exceptions (like
infrared and TPM on I945 Thinkpads) most of the hardware work fine.

However we currently need to track what computers were tested, if the
installation instruction work, and the upstream status to know if a
computer is likely to continue being supported or not if nobody
volunteers to help.

This is also a practical way to ask volunteers to test with very
simple tasks like bug reporting if things work or don't work.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>
Acked-by: Adrien 'neox' Bourmault <neox@gnu.org>