site: fix the GNU Boot build instructions.
The reference to lbmk was removed because there is no lbmk script in
GNU Boot.

The python section was removed as this is already taken care of by the
combination of the dependencies installation scripts and the
recommendation of specific distributions to build GNU Boot. Untested
distributions were also removed.

As for the removal of x-toc-enable, we had issues when having both
x-toc-enable and x-reviewed on the same page in the past. This is
documented in the commit adc578ae49e0283b0f42e34a7b2f53464705c827
("site: license.md: fix x-reviewed"). So to make sure to avoid similar
issue we preemptively remove x-toc-enable.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>
Acked-by: Adrien 'neox' Bourmault <neox@gnu.org>