site: remove docs/maintain and references to it.
This page is supposed to describe the Libreboot build system, but it
seems to only be there as a placeholder waiting for contributions that
add the actual information.

The problem here is that we don't want to describe the build system in
details in the manual for now as it is being reworked and it is
subject to change.

So for now we prefer to cleanup the code instead, and commenting it
along the way if possible, and if some explanations need to be added
in the manual it will most likely cover the global architecture which
isn't defined yet.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>
Acked-by: Adrien 'neox' Bourmault <neox@gnu.org>