site: build instructions: mention where to find binaries.
Documenting where to find binaries built with 'make release' helps
users trying to build GNU Boot for the first time, especially if
something goes wrong.

It is also extremely useful for contributors writing release scripts
as this avoid needing to build a complete release just to find out the
name and location of released files.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>
Acked-by: Adrien 'neox' Bourmault <neox@gnu.org>