Makefile: fix pipefail
In the commit d68b4f8930dc47ca23c0d88347a8c33472a9b930 ("Makefile: Add
logging.") I forgot to add the SHELL := $(shell which bash) line that
is mentionned in the commit to the Makefile.

So now the Makefile targets are actually broken:
$ make release
    echo 'Makefile: running release target' >> make-1705163885.log
    set -o pipefail ; ./build release src | tee -a make-1705163885.log
    /bin/sh: 1: set: Illegal option -o pipefail
    make: *** [Makefile:59: release] Error 2

Since 'SHELL := bash' also works on PureOS 10 and Trisquel 10, I'll
use that instead as it is more simple.

The path of bash is not hardcoded to respect users decisions to use
the bash they want, and also to work better with non-FHS
distributions.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>
Signed-off-by: Adrien 'neox' Bourmault <neox@gnu.org>