Update GRUB to 2.12.
The original patch was to bring various improvements (from its commit
message: "GRUB 2.12 brings many improvements including, but not
limited to, detached LUKS keyfile support, and numerous fixes for file
system behavior, compared to GRUB 2.06 [...]").

It was delayed as updating GRUB was not a priority at the time since
it required re-testing many computers.

However recent CVEs[1][2] can potentially break secure boot (no
exploits are available, but this is mentioned in some of the CVEs),
including freedom respecting versions that are implemented with full
disk encryption (without a /boot) or with GPG signatures and GRUB.

[1][SECURITY PATCH 00/73] GRUB2 vulnerabilities - 2025/02/18
   From Daniel Kiper (Tue, 18 Feb 2025 19:00:29 +0100)[2]
[2]https://lists.gnu.org/archive/html/grub-devel/2025-02/msg00024.html

These feature were used by some Libreboot users and also
documented/advertised in many places (GRUB manual, forums, various
wikis (Coreboot, Parabola), etc), so even if the GNU Boot manual
doesn't document them yet, the vulnerabilities should be fixed.

This change updates GRUB to 2 commits after 2.12 to get the commit
b835601c7639ed1890f2d3db91900a8506011a8e ("build: Include
grub-core/extra_deps.lst in dist") which "Fixes build failure due to
the extra_deps.lst file not existing in the tarball.".

Updating to 2.12 (+ 2 commits) will not fix the security
vulnerabilities. To do that we need to backport all the fixes
ourselves[1][2] or use the latest git revision.

This has risk, as Arch Linux "pushed a package built from git
master [but then] realized that was suffering issues and pulled it
from the repository."[1][2].

[1]Re: [SECURITY PATCH 00/73] GRUB2 vulnerabilities - 2025/02/18
   From Christian Hesse (Fri, 21 Feb 2025 11:06:54 +0100)[2].
[2]https://lists.gnu.org/archive/html/grub-devel/2025-02/msg00124.html

Because of that it is easier to update to 2.12 first to make it easier
to bisect any potential issues as at the time of writing,

At the time of writing, GRUB 2.12 or from git were tested on the
ThinkPad X200, X301, T400s, and also the ASUS KGPE-D16.

Signed-off-by: Leah Rowe <leah@libreboot.org>
GNUtoo: commit message rewrite, rebased.
Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>
Tested-by: Adrien 'neox' Bourmault <neox@gnu.org>
neox: commit message update
Acked-by: Adrien 'neox' Bourmault <neox@gnu.org>