Makefile.am: Document commands from site/docs/build/index.md.
The build documentation (site/docs/build/index.md) references some
commands, so documenting them in the a 'make help' command as well can
help improve the contributors efficiency by not having to re-read the
documentation every time we are looking for a command.

In addition this can help separate well known commands from other less
documented ones that are internal to GNU Boot and that can more easily
change.

This is also important as it allows, if the need arrise, to have the
the autotools system play a more important role than merely a
"compatibility system" for people used to type "make" as the build
instructions currently state.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>
Acked-by: Adrien 'neox' Bourmault <neox@gnu.org>