gitignore: don't ignore .scm files.
For some reason the regex being removed ends up ignoring .scm files:
if we run git add to some files in ./resources/guix/*/*.scm, we end up
with:
    The following paths are ignored by one of your .gitignore files:
    [...]
    hint: Use -f if you really want to add them.
    hint: Turn this message off by running
    hint: "git config advice.addIgnoredFile false"
but if we remove '*.s[a-w]?' from .gitignore, git add works fine
without any warning.

As this gitignore entry was added by the very first commit of this
repository (commit 89517ed6b9e89321e1387ed8f534d7f080ffdf57
"libreboot!"), we don't really know why it was added (what type of
files it's supposed to match again).

But given that we plan to use .scm files extensively, it's probably
better to remove that ignore rule and find out later the reason why it
was added in the first place, and adjust the rules accordingly, than
risking to forget some important scheme files.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>
Acked-by: Adrien 'neox' Bourmault <neox@gnu.org>