site: git.md: document research on pseudonymous contributions.
The Guix manual[1] allows patches under pseudonymous:
    Contributors are not required to use their legal name in patches
    and on-line communication; they can use any name or pseudonym of
    their choice.
[1]https://guix.gnu.org/en/manual/en/html_node/Contributing.html

For Guix packages that contain patches to other upstream projects we
would need to contact the upstream project to best see how to handle
that.

As for the documentation, part of it is most likely to stay specific
to GNU Boot, but other parts could be merged into Guix (which was
already checked), GRUB or other projects to simplify maintenance.

As for the GNU Boot build system, it's most likely specific to GNU
Boot or likely to be moved into Guix packages at some point. The only
problematic cases that can remain are the move of the build logic
inside upstream project or for transforming hacks into autotools
contributions.

But in any case when contributions that could be merged in projects
that weren't checked are made, the GNU Boot maintainers and/or
followers of the GNU Boot patches mailing list could simply mention
the issue and look up the upstream policy with regard to pseudonymous
contributions, and if that doesn't work out, work with upstream
projects to find a solution.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>
Acked-by: Adrien 'neox' Bourmault <neox@gnu.org>