website-build: use a less common port.
When working on multiple websites the same time, having every website
using the port 8080 by default isn't very convenient.

In addition here we need a number that is somewhat easy to remember
and possibly meaningful for each website, so in general that not easy
to choose.

But in the case of GNU Boot we can find a scheme because:
- GNU Boot supports only x86 machines for now,
- the GM45 ThinkPads are probably the most well supported by GNU
  Boot and also well known machines,
- that the GM45 ThinkPads use Intel CPUs,
- that 8086 is frequently used as vendor ID for Intel,

so here we can simply use 8086 as it's meaningful and at least very
easy to remember at least for people working in that field.

In addition for people not familiar with writing drivers for x86
computers, it could also reminds of a CPU also from Intel.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>
Acked-by: Adrien 'neox' Bourmault <neox@gnu.org>